+++ src/DependencyInjection/Configuration.php
@@ -12,6 +12,7 @@
 namespace Nelmio\SolariumBundle\DependencyInjection;
 
 use Solarium\Client;
+use Symfony\Component\Config\Definition\BaseNode;
 use Symfony\Component\Config\Definition\Builder\TreeBuilder;
 use Symfony\Component\Config\Definition\ConfigurationInterface;
 
@@ -26,13 +27,7 @@ class Configuration implements ConfigurationInterface
     public function getConfigTreeBuilder(): TreeBuilder
     {
         $treeBuilder = new TreeBuilder('nelmio_solarium');
-
-        if (method_exists($treeBuilder, 'getRootNode')) {
-            $rootNode = $treeBuilder->getRootNode();
-        } else {
-            // BC layer for symfony/config < 4.2
-            $rootNode = $treeBuilder->root('nelmio_solarium');
-        }
+        $rootNode = $treeBuilder->getRootNode();
 
         $rootNode
             ->children()
@@ -48,7 +43,11 @@ class Configuration implements ConfigurationInterface
                             ->scalarNode('port')->defaultValue(8983)->end()
                             ->scalarNode('path')->defaultValue('/')->end()
                             ->scalarNode('core')->end()
-                            ->scalarNode('timeout')->defaultValue(5)->end()
+                            ->scalarNode('timeout')
+                                ->setDeprecated(
+                                    ...$this->getDeprecationMsg('Configuring a timeout per endpoint is deprecated. Configure the timeout on the client adapter instead.', '4.1')
+                                )
+                            ->end()
                         ->end()
                     ->end()
                 ->end()
@@ -57,9 +56,27 @@ class Configuration implements ConfigurationInterface
                     ->useAttributeAsKey('name')
                     ->prototype('array')
                         ->addDefaultsIfNotSet()
+                        ->validate()
+                            ->ifTrue(function ($v) {
+                                return !empty($v['adapter_timeout']) && !empty($v['adapter_service']);
+                            })
+                            ->thenInvalid('Setting "adapter_timeout" is only supported for the default adapter and not in combination with "adapter_service".')
+                        ->end()
+                        ->validate()
+                            ->ifTrue(function ($v) {
+                                return !empty($v['adapter_class']) && !empty($v['adapter_service']);
+                            })
+                            ->thenInvalid('Only either "adapter_timeout" or "adapter_class" can be set')
+                        ->end()
                         ->children()
-                            ->scalarNode('client_class')->cannotBeEmpty()->defaultValue('Solarium\Client')->end()
-                            ->scalarNode('adapter_class')->end()
+                            ->scalarNode('client_class')->cannotBeEmpty()->defaultValue(Client::class)->end()
+                            ->scalarNode('adapter_class')
+                                ->setDeprecated(
+                                    ...$this->getDeprecationMsg('Configuring an adapter class is deprecated. Configure an adapter service instead.', '4.1')
+                                )
+                            ->end()
+                            ->scalarNode('adapter_timeout')->end()
+                            ->scalarNode('adapter_service')->end()
                             ->arrayNode('endpoints')
                                 ->beforeNormalization()
                                     ->ifString()
@@ -146,5 +163,28 @@ class Configuration implements ConfigurationInterface
         return function ($endpointList) {
             return preg_split('/\s*,\s*/', $endpointList);
         };
+    }
+
+    /**
+     * Returns the correct deprecation param's as an array for setDeprecated.
+     *
+     * Symfony/Config v5.1 introduces a deprecation notice when calling
+     * setDeprecation() with less than 3 args and the getDeprecation method was
+     * introduced at the same time. By checking if getDeprecation() exists,
+     * we can determine the correct param count to use when calling setDeprecated.
+     *
+     * @return string[]
+     */
+    private function getDeprecationMsg(string $message, string $version): array
+    {
+        if (method_exists(BaseNode::class, 'getDeprecation')) {
+            return [
+                'nelmio/solarium-bundle',
+                $version,
+                $message,
+            ];
+        }
+
+        return [$message];
     }
 }